+++ packages/orq-rc/RELEASES.md
@@ -6999,3 +6999,13 @@ Based on:
 - [typescript v4.4.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.17 - packages/orq-rc
+
+## 2026-02-20 11:20:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.723.0 (2.835.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.18 - packages/orq-rc